requirements.txt
fastapi
uvicorn
httpx
python-dotenv
requests
apscheduler
jinja2
google-auth==2.38.0
google-cloud-aiplatform==1.86.0
pydantic==2.6.1
google-genai==1.11.0
xxhash
openai==1.76.0

mysql-connector-python
